+++ src/parser/ast/literals/literal.rs
@@ -1,8 +1,8 @@
 #[derive(Debug, Clone, PartialEq)]
 pub enum Literal {
     String(String),
-    Int(i32),
-    Float(f32),
+    Int(i64),
+    Float(f64),
     Bool(bool),
     Null
 }
